+++ src/Root.tsx
@@ -5,82 +5,81 @@ import {
 
 import Home from "./routes/home";
 import HealthDashboard from "./routes/health-dashboard";
-import JobMonitoring from "./routes/jobs/monitoring";
-import NewJob from "./routes/jobs/new";
 
 import Navbar from "./components/Navbar"
 import WebView from "./components/WebView";
 
-import { getProcesses, getProcessRoute } from "./utils/processes";
 import NotFound from "./routes/errors/not-found";
-import { healthDataRequiresFetchOrUpdate } from "./state/selectors/healthSelectors";
 import { useAppDispatch, useAppSelector } from "./state/hooks";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { getHealthData } from "./state/slices/healthSlice";
+import { formatRoute } from "./utils/strings";
+import { Progress } from "@nasa-jpl/react-stellar";
 
 function Root() {
   
   const dispatch = useAppDispatch();
-  const processes = getProcesses();
   
   const healthState = useAppSelector((state) => {
     return state.health;
   });
-  
+
+  const [loading, setLoading] = useState(true);
+
   useEffect(() => {
-    
-    //let isMounted = true;
-    
-    // Check if data manager status is 'idle', then fetch the investigations data from the API
-    if (healthDataRequiresFetchOrUpdate(healthState)) {
+
+
+    if (healthState.status === "idle") {
+      // Fetch the health data
       dispatch(getHealthData());
-    }
-    
-    if (healthState.status === "pending") {
-      // Do something to inform user that investigation data is being fetched
+    } else if ( healthState.status === "pending" ) {
+      // Do something to inform the user that the health data is being fetched
     } else if (healthState.status === "succeeded") {
       // Do something to handle the successful fetching of data
-    } else if (healthState.error != null || healthState.error != undefined) {
+      setLoading(false);
+    } else if (healthState.status === "failed") {
       // Do something to handle the error
-      console.log(healthState.error);
+      
+      setLoading(false);
+
+      if( import.meta.env.DEV ) {
+        console.error("Health API Request Error:", healthState.error);
+      }
+
     }
-    
-    // Cleanup function
-    return () => {
-      //isMounted = false;
-    };
-    
+
   }, [healthState, dispatch]);
-  
+
   return (
-    <div className="viewWrapper">
-    <Navbar />
-    <div className="view">
-    <Routes>
+    <div className="unity-view-wrapper">
       {
-        healthState.items.map( (item, index) => {
-          return <Route key={index} path={"/applications/" + item.service} element={<WebView url={item.landingPage} />} />
-        })
+        loading && <>
+          <div className="unity-view unity-view-with-status">
+            <div className="progressStatus">
+              <h2>Setting up Unity UI Application</h2>
+              <Progress />
+            </div>
+          </div>
+        </>
       }
-      {/*<Route path="/applications/catalog" element={<WebView url={Config.ads.url} />} />*/}
-      <Route path="/health-dashboard" element={<HealthDashboard />} />
-      <Route path="/jobs/monitoring" element={<JobMonitoring />} />
-      <Route path="/jobs/monitoring/:jobid_param" element={<JobMonitoring />} />
-      <Route path="/jobs/new" element={<NewJob />} />
-      {
-        /* Add routes for job execution forms */
-        processes.map( (item) => {
-          const path = "/jobs/new/" + item['id'];
-          const route:JSX.Element | null = getProcessRoute(item['id']);
-          return (
-            <Route path={path} element={ (route) ? route : <NotFound />} key={"route_" + item['id']}/>
-          )
-        })
+      { 
+        !loading && <>
+          <Navbar />
+          <div className="unity-view unity-view-with-navbar">
+          <Routes>
+            {
+              healthState.items.map( (item, index) => {
+                return <Route key={index} path={"/applications/" + formatRoute(item.componentName)} element={<WebView url={item.landingPageUrl} />} />
+              })
+            }
+            {/*<Route path="/applications/catalog" element={<WebView url={Config.ads.url} />} />*/}
+            <Route path="/health-dashboard" element={<HealthDashboard />} />
+            <Route path="/" element={<Home />} />
+            <Route path="*" element={<NotFound />} />
+          </Routes>
+          </div>
+        </>
       }
-      <Route path="/" element={<Home />} />
-      <Route path="*" element={<NotFound />} />
-    </Routes>
-    </div>
     </div>
   )
 }