+++ src/routes/health-dashboard/index.tsx
@@ -1,14 +1,37 @@
-import { AgGridReact } from "ag-grid-react"; // the AG Grid React Component
-import { CellClickedEvent, ICellRendererParams } from 'ag-grid-community';
+import { AgGridReact, CustomCellRendererProps } from "ag-grid-react"; // the AG Grid React Component
+import { CellClickedEvent } from 'ag-grid-community';
 import { getHealthData } from "../../state/slices/healthSlice";
-import { healthDataRequiresFetchOrUpdate } from "../../state/selectors/healthSelectors";
 import { DocumentMeta } from "../../components/DocumentMeta/DocumentMeta";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useAppDispatch, useAppSelector } from "../../state/hooks";
-import { IconWarning } from "@nasa-jpl/react-stellar";
+import { IconWarning, Error } from "@nasa-jpl/react-stellar";
 import React from "react";
+import { Link } from "react-router-dom";
+
+ const LinkCellRenderer = (props:CustomCellRendererProps) => {
+  return <Link to={props.value} target="_blank">{props.value}</Link>
+}
+
+const StatusCellRenderer = (props:CustomCellRendererProps) => {
+
+  if( props.value ) {
+    let icon;
+    if( props.value.toString().toUpperCase() === "UNHEALTHY" ) {
+      icon = <IconWarning className="unity-icon-warning"/>
+    }
+    if( props.value.toString().toUpperCase() === "UNAVAILABLE" ) {
+      icon = <IconWarning className="unity-icon-error"/>
+    }
+
+    return <React.Fragment>{props.value} {icon}</React.Fragment>
+  }
+
+}
 
 function HealthDashboard() {
+
+  const [healthApiError, setHealthApiError] = useState(false);
+  const healthApiErrorMessage = "Health Endpoint is Not Available";
 
   const dispatch = useAppDispatch();
 
@@ -18,30 +41,23 @@ function HealthDashboard() {
 
   // Each Column Definition results in one Column.
   const [columnDefs] = useState([
-    { field: "service", headerName: "Service", filter: true },
+    { field: "componentName", headerName: "Service", filter: true },
     {
       cellClass: 'unity-aggrid-health-status',
-      cellRenderer: (params:ICellRendererParams) => {
-
-        if( params.value ) {
-          let icon;
-          if( params.value.toString().toUpperCase() === "UNHEALTHY" ) {
-            icon = <IconWarning className="unity-icon-warning"/>
-          }
-          if( params.value.toString().toUpperCase() === "UNAVAILABLE" ) {
-            icon = <IconWarning className="unity-icon-error"/>
-          }
-  
-          return <React.Fragment>{params.value} {icon}</React.Fragment>
-        }
-
-      },
+      cellRenderer: StatusCellRenderer,
       field: "status", 
       filter: true,
       headerName: "Status",
       valueGetter: "data.healthChecks[0].status",
     },
-    { field: "landingPage", headerName: "Landing Page", filter: true, cellStyle: { cursor: 'pointer', color: '#0000FF', textDecoration: 'underline' }},
+    { field: "landingPageUrl", 
+      cellRenderer: LinkCellRenderer,
+      headerName: "Landing Page", filter: true, cellStyle: { color: '#0000FF', textDecoration: 'underline' }
+    },
+    { field: "healthCheckUrl", 
+      cellRenderer: LinkCellRenderer,
+      headerName: "Health Check Page", filter: true, cellStyle: { color: '#0000FF', textDecoration: 'underline' }
+    },
     { field: "date", valueGetter: "data.healthChecks[0].date", headerName: "Date", filter: true },
   ]);
 
@@ -56,32 +72,41 @@ function HealthDashboard() {
   // Example of consuming Grid Event
   const cellClickedListener = useCallback( (event:CellClickedEvent) => {
 
-    if( event.colDef.field === 'landingPage') {
-      window.location.href = "https://" + event.data.landingPage;
+    if( import.meta.env.DEV ) {
+      // Output information to help with debugging only in DEV mode
+      console.info(`The field ${event.colDef.field} containing the data, ${event.data} was clicked.`);
     }
+
+    // Example of how to 
+    /*if( event.colDef.field === 'landingPageUrl') {
+      window.location.href = event.data.landingPageUrl;
+    }*/
 
  }, []);
 
   const onGridReady = useCallback( () => {
     const abortController = new AbortController();
 
-    // Check if data manager status is 'idle', then fetch the investigations data from the API
-    if (healthDataRequiresFetchOrUpdate(healthState)) {
+    // Fetch the health data
+    if (healthState.status === "idle") {
+      // Fetch the health data
       dispatch(getHealthData());
-    }
+    } 
     
     return () => abortController.abort();
   }, [dispatch, healthState]);
 
   useEffect(() => {
+
     //let isMounted = true;
 
-    if (healthState.status === "pending") {
-      // Do something to inform user that investigation data is being fetched
+    if ( healthState.status === "pending" ) {
+      // Do something to inform the user that the health data is being fetched
     } else if (healthState.status === "succeeded") {
       // Do something to handle the successful fetching of data
-    } else if (healthState.error != null || healthState.error != undefined) {
+    } else if (healthState.status === "failed") {
       // Do something to handle the error
+      setHealthApiError(true);
       console.log(healthState.error);
     }
 
@@ -96,7 +121,8 @@ function HealthDashboard() {
       <DocumentMeta title="Health Dashboard" description="Health Dashboard" />
       <div className="main-view">
         <h1>Health Dashboard</h1>
-        <div className="ag-theme-stellar data-grid-container">
+        { healthApiError && <Error><IconWarning />{healthApiErrorMessage}</Error> }
+        <div className="ag-theme-stellar unity-aggrid-container">
           <AgGridReact
             rowData={healthState.items} // Row Data for Rows
             columnDefs={columnDefs} // Column Defs for Columns