+++ src/routes/home/index.tsx
@@ -1,6 +1,10 @@
 import { DocumentMeta } from "../../components/DocumentMeta/DocumentMeta"
+import Config from "../../Config";
 
 function Home() {
+
+  const project = Config['general']['project'];
+  const venue = Config['general']['venue'];
 
    return (
       <>
@@ -10,6 +14,8 @@ function Home() {
          />
          <div className="main-view">
             <h1>Home</h1>
+            <div>Project: <strong>{project}</strong></div>
+            <div>Venue: <strong>{venue}</strong></div>
          </div>
       </>
    )