+++ src/state/slices/healthSlice.ts
@@ -1,5 +1,7 @@
 import axios, { AxiosRequestConfig } from 'axios';
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+import Config from "../../Config";
+import { GetToken } from '../../AuthorizationWrapper';
 
 enum HEALTH_ACTIONS {
   GET_HEALTH = "health/getHealth",
@@ -7,13 +9,16 @@ enum HEALTH_ACTIONS {
 
 type HealthCheck = {
   status:string;
+  httpResponseCode:string;
   date:string | number;
 }
 
 export type Service = {
-  service:string;
-  landingPage:string;
-  healthChecks: Array<HealthCheck>
+  componentName:string;
+  healthCheckUrl:string;
+  landingPageUrl:string;
+  healthChecks: Array<HealthCheck>;
+  ssmKey:string;
 };
 
 export type HealthState = {
@@ -30,6 +35,96 @@ const initialState:HealthState = {
   status: 'idle',
 };
 
+const getItems = () => {
+
+  // Get additional links that should be added for the given project and venue
+
+  const project = Config['general']['project'].toUpperCase();
+  const venue = Config['general']['venue'].toUpperCase();
+
+  let serviceItems:Service[] = Array<Service>();
+
+  if( project === "UNITY" && venue === 'OPS') {
+
+    serviceItems = [
+      {
+        componentName: "STAC Browser",
+        ssmKey: "",
+        healthCheckUrl: "",
+        landingPageUrl: "https://www.mdps.mcp.nasa.gov:4443/data/stac_browser/",
+        healthChecks: [
+          {
+            status: "UNKNOWN",
+            httpResponseCode: "",
+            date: ""
+          }
+        ]
+      }
+    ];
+  }
+
+  if( project === "EMIT" && venue === "DEV" ) {
+
+    serviceItems = [
+      {
+        componentName: "Jupyterhub",
+        ssmKey: "",
+        healthCheckUrl: "",
+        landingPageUrl: "https://www.mdps.mcp.nasa.gov:4443/emit/dev/jupyter/",
+        healthChecks: [
+          {
+            status: "UNKNOWN",
+            httpResponseCode: "",
+            date: ""
+          }
+        ]
+      },
+      {
+        componentName: "Airflow",
+        ssmKey: "",
+        healthCheckUrl: "",
+        landingPageUrl: "http://k8s-sps-airflowi-9a4fb23ed9-117303406.us-west-2.elb.amazonaws.com:5000/",
+        healthChecks: [
+          {
+            status: "UNKNOWN",
+            httpResponseCode: "",
+            date: ""
+          }
+        ]
+      },
+      {
+        componentName: "Airflow-ogc",
+        ssmKey: "",
+        healthCheckUrl: "",
+        landingPageUrl: "http://k8s-sps-ogcproce-927cdf8d63-717063809.us-west-2.elb.amazonaws.com:5001/",
+        healthChecks: [
+          {
+            status: "UNKNOWN",
+            httpResponseCode: "",
+            date: ""
+          }
+        ]
+      },
+      {
+        componentName: "STAC Browser",
+        ssmKey: "",
+        healthCheckUrl: "",
+        landingPageUrl: "https://www.mdps.mcp.nasa.gov:4443/data/stac_browser/",
+        healthChecks: [
+          {
+            status: "UNKNOWN",
+            httpResponseCode: "",
+            date: ""
+          }
+        ]
+      }
+    ];
+  }
+
+  return serviceItems;
+
+}
+
 /**
  * Get all the instruments from the PDS OpenSearch API
  */ 
@@ -37,19 +132,19 @@ export const getHealthData = createAsyncThunk(
   HEALTH_ACTIONS.GET_HEALTH,
   async (_:void, thunkAPI) => {
     
-    // todo: fill in URL in the future when an API is available to fetch
-    // health JSON
-    const url = "/data/health.json"
+    const url = Config['cs']['health_endpoint'];
+    const token = GetToken();
     
     const config:AxiosRequestConfig = {
       headers: {
-        "Content-Type": "application/json"
+        "Content-Type": "application/json",
+        "Authorization": "Bearer " + token
       }
     }
     
     try {
       const response = await axios.get(url, config);
-      return response.data;
+      return response.data.services;
     } catch (err:any) {
       return thunkAPI.rejectWithValue({ error: err.message });
     }
@@ -62,7 +157,7 @@ const healthSlice = createSlice({
   initialState,
   reducers: {},
   extraReducers: (builder) => {
-    
+
     builder.addCase(getHealthData.pending, (state, _action) => {
       // When data is being fetched
       state.status = "pending";
@@ -75,14 +170,15 @@ const healthSlice = createSlice({
 
       // Parse and store the fetched data into the state
       const data = action.payload;
-      state.items = data;
+      state.items = data.concat(getItems());
 
     });
     
     builder.addCase(getHealthData.rejected, (state, action) => {
       // When data is fetched unsuccessfully
       state.status = "failed";
-      
+
+      state.items = getItems();
       // Update the error message for proper error handling
       state.error = action.error.message;
     });