+++ vite.config.ts
@@ -4,16 +4,12 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   envDir: ".env",
+  build: {
+    minify: false
+  },
   plugins: [
     react()
   ],
-  resolve: {
-    alias: {
-      crypto: 'crypto-browserify',
-      stream: "stream-browserify",
-      util: 'util/'
-    }
-  },
   server: {
     port: 8080
   }